requirements/local.txt
-r base.txt

pytest==7.2.2
pytest-django==4.5.2

factory-boy==3.2.1
Faker==17.6.0

# ipdb==0.13.9
# ipython==8.6.0

django-debug-toolbar==3.8.1

# mypy==0.991

# django-stubs==1.14.0
# djangorestframework-stubs==1.8.0
# boto3-stubs==1.26.41

# flake8==6.0.0
# isort==5.10.1
# black==22.10.0
# pre-commit==3.0.2
